% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independence.R
\name{independence}
\alias{independence}
\title{An independence test using R}
\usage{
independence(d1, d2, alpha)
}
\arguments{
\item{d1}{the first data vector}

\item{d2}{the second data vector}

\item{alpha}{the significance for test}
}
\value{
p the true significance

T the result of the test
}
\description{
An independence test using R by the nonparametric method.
}
\examples{
\dontrun{
d1<-faithful$eruptions
d2<-faithful$waiting
Test<-independence(d1,d2,0.05)
Test$T
Test$p
}
}
