% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MetropolisC}
\alias{MetropolisC}
\title{MetropolisC}
\usage{
MetropolisC(sigma, x0, N)
}
\arguments{
\item{sigma}{the variance of the normal}

\item{x0}{the initial value}

\item{N}{the number of the trying}
}
\value{
the path and the number of the accept point
}
\description{
The Rcpp function in homework.
}
\examples{
\dontrun{
x<-Metropolisc(1,25,2000)
x
}
}
