\name{SC19025-package}
\alias{SC19025-package}
\alias{SC19025}
\docType{package}
\title{
  The final homework for 'Statistical Computing' course
}
\description{
  The functions for nonparametric hypothesis testing and all my homework.
}

\author{
SC19025

Maintainer: SC19025 <2018773571@qq.com>
}

\keyword{ package }
