% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetry.R
\name{symmetry}
\alias{symmetry}
\title{A symmetry test using R}
\usage{
symmetry(d, alpha)
}
\arguments{
\item{d}{the data vector}

\item{alpha}{the significance for test}
}
\value{
p the true significance

T the result of the test
}
\description{
A symmetry test using R by the nonparametric method.
}
\examples{
\dontrun{
d1<-rnorm(100)
d2<-rchisq(100,df=1)
Test1<-symmetry(d1,0.05)
Test1
Test2<-symmetry(d2,0.05)
Test2
}
}
